XC-331: move to a fixture file
