+++ libs/client/src/lib.rs
@@ -38,6 +38,8 @@
 #![forbid(unsafe_code)]
 #![forbid(missing_docs)]
 
+#[cfg(not(target_arch = "wasm32"))]
+pub mod fixtures;
 mod request;
 
 pub use request::{Request, RequestBuilder, SolRpcEndpoint, SolRpcRequest};
@@ -48,8 +50,8 @@ use candid::{utils::ArgumentEncoder, CandidType, Principal};
 use ic_cdk::api::call::RejectionCode;
 use serde::de::DeserializeOwned;
 use sol_rpc_types::{
-    GetAccountInfoParams, GetSlotParams, GetSlotRpcConfig, MultiRpcResult, RpcConfig, RpcSources,
-    SolanaCluster, SupportedRpcProvider, SupportedRpcProviderId,
+    GetAccountInfoParams, GetSlotParams, GetSlotRpcConfig, RpcConfig, RpcSources, SolanaCluster,
+    SupportedRpcProvider, SupportedRpcProviderId,
 };
 use solana_clock::Slot;
 use std::sync::Arc;
@@ -165,16 +167,6 @@ impl<R> ClientBuilder<R> {
                 rpc_sources: self.config.rpc_sources,
             },
         }
-    }
-
-    /// Change the runtime to return a mocked response.
-    ///
-    /// *IMPORTANT*: This method should only be used for testing purposes.
-    pub fn with_mocked_response<Out: CandidType>(
-        self,
-        mocked_response: Out,
-    ) -> ClientBuilder<MockRuntime> {
-        self.with_runtime(|_runtime| MockRuntime::new(mocked_response))
     }
 
     /// Mutates the builder to use the given [`RpcSources`].
@@ -389,52 +381,5 @@ impl Runtime for IcRuntime {
         ic_cdk::api::call::call(id, method, args)
             .await
             .map(|(res,)| res)
-    }
-}
-
-/// A dummy implementation of [`Runtime`] that always return the same response.
-pub struct MockRuntime(Vec<u8>);
-
-impl MockRuntime {
-    /// Create a new [`MockRuntime`] to always return the given parameter.
-    pub fn new<Out: CandidType>(mocked_response: Out) -> Self {
-        Self(
-            candid::encode_args((&mocked_response,))
-                .expect("Failed to encode Candid mocked response"),
-        )
-    }
-}
-
-#[async_trait]
-impl Runtime for MockRuntime {
-    async fn update_call<In, Out>(
-        &self,
-        id: Principal,
-        method: &str,
-        args: In,
-        cycles: u128,
-    ) -> Result<Out, (RejectionCode, String)>
-    where
-        In: ArgumentEncoder + Send,
-        Out: CandidType + DeserializeOwned,
-    {
-        Ok(candid::decode_args(&self.0)
-            .map(|(r,)| r)
-            .expect("Failed to decode Candid mocked response"))
-    }
-
-    async fn query_call<In, Out>(
-        &self,
-        id: Principal,
-        method: &str,
-        args: In,
-    ) -> Result<Out, (RejectionCode, String)>
-    where
-        In: ArgumentEncoder + Send,
-        Out: CandidType + DeserializeOwned,
-    {
-        Ok(candid::decode_args(&self.0)
-            .map(|(r,)| r)
-            .expect("Failed to decode Candid mocked response"))
     }
 }
